restructing the files to implement mern structure. separated graphql functions in a separate folder
