+++ graphql/resolvers/posts.js
@@ -1,0 +1,14 @@
+const Post = require('../../models/Post');
+
+module.exports = {
+        Query: {
+            async getPosts(){
+                try{
+                    const posts = await Post.find();
+                    return posts;
+                } catch(err) {
+                    throw new Error(err);
+                }
+            }
+        }
+}
