+++ index.js
@@ -1,35 +1,11 @@
 const { ApolloServer } = require('apollo-server');
-const gql = require('graphql-tag'); 
 const mongoose = require('mongoose');
 
-const Post = require('./models/Post');
+const typeDefs = require('./graphql/typeDefs');
+const resolvers = require('./graphql/resolvers');
 const { MONGODB } = require('./config.js');
 
 
-const typeDefs = gql`
-    type Post{
-        id: ID!
-        body: String!
-        createdAt: String!
-        username: String!
-    }
-    type Query {
-        getPosts: [Post]
-    }
-`;
-
-const resolvers = {
-    Query: {
-        async getPosts(){
-            try{
-                const posts = await Post.find();
-                return posts;
-            } catch(err) {
-                throw new Error(err);
-            }
-        }
-    }
-}
 
 const server = new ApolloServer({
     typeDefs,
