+++ package.json
@@ -1,0 +1,27 @@
+{
+  "name": "oneness",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "dependencies": {
+    "apollo-server": "^2.20.0",
+    "graphql": "^15.5.0",
+    "mongoose": "^5.11.15",
+    "nodemon": "^2.0.7"
+  },
+  "devDependencies": {},
+  "scripts": {
+    "start": "nodemon index"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/samergain/oneness.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/samergain/oneness/issues"
+  },
+  "homepage": "https://github.com/samergain/oneness#readme"
+}
